requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
langdetect==1.0.9
sentence-transformers==3.2.1
faiss-cpu==1.9.0
hyperon==0.2.8
pydantic~=2.12.0
python-dotenv==1.0.1
loguru==0.7.2
pytest==8.3.3
httpx~=0.28.1
black==24.10.0

click~=8.2.1
h11~=0.16.0
numpy~=2.3.2
pillow~=11.3.0
pip~=24.3.1
Mako~=1.3.10
MarkupSafe~=3.0.3
mypy~=1.13.0
typing_extensions~=4.15.0
mypy_extensions~=1.1.0
python-telegram-bot~=22.3
sniffio~=1.3.1
platformdirs~=4.5.0
packaging~=25.0
pathspec~=0.12.1
httpcore~=1.0.9
idna~=3.10
certifi~=2025.8.3
charset-normalizer~=3.4.3
regex~=2024.9.11
matplotlib~=3.10.6
tqdm~=4.67.1
flake8~=7.1.1
six~=1.17.0
pyflakes~=3.2.0
pycodestyle~=2.12.1
joblib~=1.5.2
pluggy~=1.6.0
iniconfig~=2.3.0
alembic~=1.13.3
starlette~=0.40.0
pydantic_core~=2.41.1
anyio~=4.10.0
itsdangerous~=2.2.0
Jinja2~=3.1.6
nltk~=3.9.1
annotated-types~=0.7.0
typing_extensions~=4.15.0
mypy_extensions~=1.1.0
pydantic_core~=2.41.1